requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
langchain==0.1.13
langchain-openai==0.0.8
openai==1.17.0
psycopg2-binary==2.9.9
pgvector==0.2.4
SQLAlchemy==2.0.28
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
faiss-cpu==1.8.0
sentence-transformers==2.5.1
